client: add support for "stdin" from snapctl
This adds initial support for having "snapctl" commands read
from "stdin". Note that in the future we want something more
advanced that keeps the connection open and forwards the real
stdin to the server.

This initial version will just read stdin for snapctl commands
that are known to read from stdin like `snapctl fde-setup-result`.