devicestate: hack around devicestate.BootSuccessful() issue
The trustedContentObserver will call "resealKeyToModeenv()"
during early seeding. At this point there is no kernel installed
yet so the standard way to check for the fde-setup hook will
not work. This commit hacks around this by checking the sealed-keys
file content. Note that this is also not great because
resealKeyToModeenvUsingFDESetupHook() will not have access to
the kernel yet in many situations.

E.g. during firstboot seeding.