snapstate: adjust update-gadget-assets user visible message
The `update-gadget-assets` task used to have "gadget" hardcoded
in the human readable description.

This commit changes the description to read
```
Update assets from gadget "<gadget-snap>"
Update assets from kernel "<kernel-snap>"
```
depending on the type.

Note that the task kind does not change, I think that's okay
because even though the assets come from the kernel snap they
are defined in the gadget.